Create a separate mod for BallistaConfig
