+++ ballista/scheduler/src/flight_sql.rs
@@ -42,7 +42,7 @@ use crate::scheduler_server::SchedulerServer;
 use arrow_flight::flight_service_client::FlightServiceClient;
 use arrow_flight::sql::ProstMessageExt;
 use arrow_flight::SchemaAsIpc;
-use ballista_core::config::BallistaConfig;
+use ballista_core::config::query::BallistaConfig;
 use ballista_core::serde::protobuf;
 use ballista_core::serde::protobuf::action::ActionType::FetchPartition;
 use ballista_core::serde::protobuf::job_status;
